fix: add authentication, fix validation retries, type mismatches, transaction safety
- Candidates service: separate public/authenticated/admin route groups
  with Frame JWT authentication middleware
- ValidateHandler: flag variant instead of retrying forever when AI
  returns unparseable JSON
- DedupHandler: mark duplicate variants as flagged instead of leaving
  them at stage=raw forever
- API service: gate AutoMigrate behind DO_DATABASE_MIGRATE env var,
  respect readOnly parameter in dbFn
- TruncateCanonicals: wrap three DELETEs in a GORM transaction
- CandidateProfile: change SalaryMin/SalaryMax from float32 to float64
  to match CanonicalJob types
- ReduceCrawlFrequency: cap at 7 days (604800s) to prevent unbounded
  interval growth
- SourceQualityHandler: require source to be paused before disabling,
  pause first if not already paused

+++ pkg/domain/candidate.go
@@ -69,8 +69,8 @@ type CandidateProfile struct {
 	PreferredLocations  string  `gorm:"type:text" json:"preferred_locations"`
 	PreferredCountries  string  `gorm:"type:text" json:"preferred_countries"`
 	RemotePreference    string  `gorm:"type:varchar(20)" json:"remote_preference"`
-	SalaryMin           float32 `gorm:"type:real" json:"salary_min"`
-	SalaryMax           float32 `gorm:"type:real" json:"salary_max"`
+	SalaryMin           float64 `gorm:"type:real" json:"salary_min"`
+	SalaryMax           float64 `gorm:"type:real" json:"salary_max"`
 	Currency            string  `gorm:"type:varchar(10)" json:"currency"`
 
 	// Additional profile fields
